+++ scanpipe/admin.py
@@ -20,20 +20,31 @@
 # ScanCode.io is a free software code scanning tool from nexB Inc. and others.
 # Visit https://github.com/nexB/scancode.io for support and download.
 
+import json
+
+from django import forms
 from django.contrib import admin
 from django.contrib.admin.utils import unquote
 from django.contrib.admin.views.main import ChangeList
+from django.core.serializers.json import DjangoJSONEncoder
 from django.http import FileResponse
 from django.http import Http404
 from django.http import QueryDict
+from django.http import StreamingHttpResponse
+from django.shortcuts import redirect
 from django.urls import path
 from django.urls import reverse
 from django.utils.html import format_html
+from django.utils.http import urlencode
 from django.utils.safestring import mark_safe
+from django.views.generic.edit import FormView
 
+from scanpipe.api.serializers import get_model_serializer
+from scanpipe.api.serializers import get_serializer_fields
 from scanpipe.models import CodebaseResource
 from scanpipe.models import DiscoveredPackage
 from scanpipe.models import ProjectError
+from scanpipe.pipes.outputs import queryset_to_csv_stream
 
 
 class ListDisplayField:
@@ -130,13 +141,91 @@ class PathListFilter(admin.SimpleListFilter):
             return queryset.filter(path__startswith=self.value())
 
 
+class Echo:
+    """
+    An object that implements just the write method of the file-like interface.
+    """
+
+    def write(self, value):
+        """
+        Write the value by returning it, instead of storing in a buffer.
+        """
+        return value
+
+
+class ExportConfigurationForm(forms.Form):
+    """
+    Configuration form for exporting data including selected fields.
+    """
+
+    include_fields = forms.MultipleChoiceField(
+        label="", widget=forms.CheckboxSelectMultiple(attrs={"checked": "checked"})
+    )
+    pks = forms.CharField(
+        widget=forms.widgets.HiddenInput,
+    )
+
+    def __init__(self, model_class, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields["include_fields"].choices = [
+            (field, field) for field in get_serializer_fields(model_class)
+        ]
+
+
+class AdminExportView(FormView):
+    """
+    A view to configure an export and download the data.
+    """
+
+    template_name = "admin/export.html"
+    form_class = ExportConfigurationForm
+    model_admin = None
+    model_class = None
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        self.model_class = self.model_admin.model
+
+    def get_form_kwargs(self):
+        kwargs = super().get_form_kwargs()
+        kwargs["model_class"] = self.model_class
+        return kwargs
+
+    def get_initial(self):
+        initial = super().get_initial()
+        initial.update({"pks": self.request.GET.get("pks")})
+        return initial
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        extra_context = {
+            "title": "Export to CSV",
+            "opts": self.model_class._meta,
+            "has_view_permission": self.model_admin.has_view_permission(self.request),
+            "media": self.model_admin.media,
+        }
+        return {**context, **extra_context}
+
+    def form_valid(self, form):
+        model_name = self.model_class._meta.model_name
+        pks = form.cleaned_data["pks"].split(",")
+        queryset = self.model_class.objects.filter(pk__in=pks)
+        fieldnames = form.cleaned_data["include_fields"]
+        output_stream = Echo()
+
+        streaming_content = queryset_to_csv_stream(queryset, fieldnames, output_stream)
+        response = StreamingHttpResponse(streaming_content, content_type="text/csv")
+        response["Content-Disposition"] = f'attachment; filename="{model_name}.csv"'
+        return response
+
+
 class ProjectRelatedModelAdmin(admin.ModelAdmin):
     """
     Regroup the common ModelAdmin values for Project related models.
     """
 
     list_select_related = True
-    actions_on_top = False
+    actions_on_top = True
     actions_on_bottom = True
     prefetch_related = None
 
@@ -161,6 +250,43 @@ class ProjectRelatedModelAdmin(admin.ModelAdmin):
 
     project_filter.short_description = "Project"
     project_filter.admin_order_field = "project"
+
+    def get_urls(self):
+        urls = []
+        actions = getattr(self, "actions", [])
+
+        if "export_to_csv" in actions:
+            opts = self.model._meta
+            export_view = AdminExportView.as_view(model_admin=self)
+            urls.append(
+                path(
+                    "export/",
+                    self.admin_site.admin_view(export_view),
+                    name=f"{opts.app_label}_{opts.model_name}_export",
+                ),
+            )
+        return urls + super().get_urls()
+
+    def export_to_csv(self, request, queryset):
+        opts = self.model._meta
+        view_url = reverse(f"admin:{opts.app_label}_{opts.model_name}_export")
+        selected = queryset.values_list("pk", flat=True)
+        params = urlencode({"pks": ",".join(str(pk) for pk in selected)})
+        return redirect(f"{view_url}?{params}")
+
+    export_to_csv.short_description = "Export selected objects to CSV"
+
+    def export_to_json(self, request, queryset):
+        model_name = queryset.model._meta.model_name
+        serializer_class = get_model_serializer(queryset.model)
+        serializer = serializer_class(queryset, many=True)
+        json_data = json.dumps(serializer.data, indent=2, cls=DjangoJSONEncoder)
+
+        response = StreamingHttpResponse(json_data, content_type="application/json")
+        response["Content-Disposition"] = f'attachment; filename="{model_name}.json"'
+        return response
+
+    export_to_json.short_description = "Export selected objects to JSON"
 
 
 def get_admin_url(obj, view="change"):
@@ -194,6 +320,7 @@ class CodebaseResourceAdmin(ProjectRelatedModelAdmin):
     search_fields = ("path", "mime_type", "file_type")
     ordering = ["path"]
     prefetch_related = ["discovered_packages"]
+    actions = ["export_to_csv", "export_to_json"]
 
     def path_filter(self, obj):
         """
@@ -211,7 +338,10 @@ class CodebaseResourceAdmin(ProjectRelatedModelAdmin):
             if last_segment:
                 links.append(f'<b><a href="{get_admin_url(obj)}">{segment}</a></b>')
             else:
-                links.append(f'<a href="?path={current_path}">{segment}</a>')
+                request = getattr(obj, "_request")
+                query_dict = request.GET.copy() if request else QueryDict()
+                query_dict["path"] = current_path
+                links.append(f'<a href="?{query_dict.urlencode()}">{segment}</a>')
 
         return mark_safe('<span class="path_separator">/</span>'.join(links))
 
@@ -287,6 +417,7 @@ class DiscoveredPackageAdmin(ProjectRelatedModelAdmin):
     exclude = ("codebase_resources",)
     inlines = (CodebaseResourceInline,)
     prefetch_related = ["codebase_resources"]
+    actions = ["export_to_csv", "export_to_json"]
 
     def resources(self, obj):
         return mark_safe(